port definitions for arrival curves
This is a port (+additions) of the definitions and semantics for arrival curves (model/arrival/curves.v). As a prerequisite, this includes definitions about activations of a task (model/task_arrivals.v).

Two additional definitions which were not found in the original library but will be useful to us in the future:

* in schedule.v : completes_at
* in task_arrivals.v : arrivals_come_from_taskset