define [backlogged] in terms of job readiness
Add a [job_ready] parameter to the job model, akin to [pending], that
defines whether a job can be scheduled at a given time in a given schedule.

Then use this new general notion of readiness to define [backlogged].

A pending job may not be ready, whereas (under any reasonable
definition) a ready job ought to be pending, so this definition is
more precise and captures effects such as jitter and self-suspensions.