add a `package.json` file to turn Prosa into an `esy` package.
`esy` is an `npm`-like package manager and build tool designed for
reproducibility, by preventing the use of unlisted dependencies and
pinning the exact version of each of these (a bit like `opam`, but
with more guarantees).

A nice feature of `esy` is ease of installation:

1. Install `esy` through `npm`. (However, note that `esy` is not a
   JavaScript program; the `esy` project just uses `npm` as a code
   hoster.

2. Run `esy` in the Prosa repository. 

This will download and compile all dependencies and compile the
repository. (On the flip side, this also means that Coq will be
compiled, so the first time, compilation will be slow.)

See the README.md file for usage instructions. 

See also the discussion: https://gitlab.mpi-sws.org/RT-PROOFS/rt-proofs/-/issues/80

For more information about `esy`, see: https://esy.sh/