CI: simplify the setup to use a single stage
We've had problems with spurious CI failures and unexpected recompilation of .vo files. Somehow the collection of .vo files in the build stage for reuse in the processing stage is not working properly. This MR simplifies the CI setup to compile, build the documentation, and validate all in one job.

It also disables builds with development versions of Coq and mathcomp (nobody was looking at this output anyway).