generalize abstract RTA
Changes:

* This commit generalises aRTA to "multi-stage" aRTA. The general idea
is explained in file [analysis/abstract/abstract_rta.v]. Short idea is
that the fixpoint equation can be extended to a sequence of fixpoint
equations. Solution to an equation can be used in later fixpoints. This
way one can support move expressive models of execution. The prior
version of aRTA (applicable to ideal uni-processors) is now an
instantiation of the new RTA theorem.

* Interference and Interfering workload are now type-classes. Note that
this changes files in directory [results], since [Variables] were
replaced with [Instance] declarations.

* IBF supports arbitrary parameters (not just relative arrival time).
IBF parametrised by the relative arrival time is not expressive enough
to support restricted-supply analysis. It was generalized to support a
larger class of parameters

* Rename [run to completion] file into [lower bound on service]. The
file was slightly generalized, now  it derives a lower bound on any
pre-defined amount of service.

* Generalize instantiations of interference and interfering workload.
Now, the definitions do not directly destruct the processor state.
Also, definitions for these functions have been made opaque.