generalize FP ⊂ GEL proof to any self-suspension model
The proof of conditional inclusion of FP in GEL does not depend on
whether jobs self-suspend or not. However, the proof so far was
limited to a specific readiness model that does not allow for
self-suspensions. This patch weakens the assumption to allow for any
readiness model (including self-suspending ones) that ensure that
tasks execute sequentially.