CI: replace `make validate` with `Print Assumptions`
Since `coqchk` is not precise, we keep running into false positives
when packages that we (indirectly) depend on (in particular,`coq-elpi`)
load the standard library. Currently, this blocks the upgrade to Coq 8.20.

This patch disables `make validate` and instead introduces a check
that runs `Print Assumptions` for each theorem in the `prosa.results`
module, which are all the main results we care about.